Move include files into include folder
instead of src/include
