+++ src/lddc.h
@@ -25,8 +25,8 @@
 #ifndef LIVOX_ROS_DRIVER2_LDDC_H_
 #define LIVOX_ROS_DRIVER2_LDDC_H_
 
-#include "include/livox_ros_driver2.h"
-#include "include/livox_ros_types.h"
+#include "livox_ros_driver2/livox_ros_driver2.h"
+#include "livox_ros_driver2/livox_ros_types.h"
 
 #include "driver_node.h"
 #include "lds.h"
