+++ src/livox_ros_driver2.cpp
@@ -28,8 +28,8 @@
 #include <csignal>
 #include <thread>
 
-#include "include/livox_ros_driver2.h"
-#include "include/ros_headers.h"
+#include "livox_ros_driver2/livox_ros_driver2.h"
+#include "livox_ros_driver2/ros_headers.h"
 #include "driver_node.h"
 #include "lddc.h"
 #include "lds_lidar.h"
